requirements.txt
awscli==1.11.183
backcall==0.1.0
bleach==2.1.3
boto==2.48.0
boto3==1.7.11
botocore==1.10.11
bz2file==0.98
cachetools==2.0.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.7
cycler==0.10.0
cymem==1.31.2
Cython==0.28.2
cytoolz==0.8.2
de-core-news-sm==2.0.0
decorator==4.3.0
dill==0.2.7.1
docutils==0.14
entrypoints==0.2.3
Flask==1.0.2
ftfy==4.4.3
gensim==3.4.0
html5lib==1.0.1
idna==2.6
ijson==2.3
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numpy==1.14.3
pandocfilters==1.4.2
parso==0.2.0
pathlib==1.0.1
pexpect==4.5.0
pickleshare==0.7.4
plac==0.9.6
preshed==1.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyemd==0.5.1
Pygments==2.2.0
pymongo==3.6.1
pyparsing==2.2.0
Pyphen==0.9.4
python-dateutil==2.7.2
python-Levenshtein==0.12.0
pytz==2018.4
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
regex==2017.4.5
requests==2.18.4
s3transfer==0.1.11
scikit-learn==0.19.1
scipy==1.0.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.5.7
spacy==2.0.11
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
textacy==0.6.1
thinc==6.10.2
toolz==0.9.0
tornado==5.0.2
tqdm==4.23.1
traitlets==4.3.2
ujson==1.35
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
-e git+https://github.com/dmlc/xgboost@9a81c74a7bd8db06c2f2ccb28b88ffeecf4c0309#egg=xgboost&subdirectory=python-package
